Moved report transform functions to offst-proto crate.
