+++ components/funder/src/report.rs
@@ -27,12 +27,6 @@ use crate::token_channel::{TokenChannel, TcDirection, TcMutation};
 use crate::liveness::LivenessMutation;
 use crate::ephemeral::{Ephemeral, EphemeralMutation};
 
-#[allow(unused)]
-#[derive(Debug)]
-pub enum ReportMutateError {
-    FriendDoesNotExist,
-    FriendAlreadyExists,
-}
 
 
 impl<A> Into<SentLocalAddressReport<A>> for &SentLocalAddress<A> 
@@ -332,193 +326,3 @@ pub fn ephemeral_mutation_to_report_mutations<A: Clone>(ephemeral_mutation: &Eph
     }
 }
 
-pub fn friend_report_mutate<A>(friend_report: &mut FriendReport<A>, mutation: &FriendReportMutation<A>) 
-where   
-    A: Clone,
-{
-    match mutation {
-        FriendReportMutation::SetRemoteAddress(remote_address) => {
-            friend_report.remote_address = remote_address.clone();
-        },
-        FriendReportMutation::SetName(name) => {
-            friend_report.name = name.clone();
-        },
-        FriendReportMutation::SetSentLocalAddress(sent_local_address_report) => {
-            friend_report.sent_local_address = sent_local_address_report.clone();
-        },
-        FriendReportMutation::SetChannelStatus(channel_status_report) => {
-            friend_report.channel_status = channel_status_report.clone();
-        },
-        FriendReportMutation::SetWantedRemoteMaxDebt(wanted_remote_max_debt) => {
-            friend_report.wanted_remote_max_debt = *wanted_remote_max_debt;
-        },
-        FriendReportMutation::SetWantedLocalRequestsStatus(wanted_local_requests_status) => {
-            friend_report.wanted_local_requests_status = wanted_local_requests_status.clone();
-        },
-        FriendReportMutation::SetNumPendingResponses(num_pending_responses) => {
-            friend_report.num_pending_responses = *num_pending_responses;
-        },
-        FriendReportMutation::SetNumPendingRequests(num_pending_requests) => {
-            friend_report.num_pending_requests = *num_pending_requests;
-        },
-        FriendReportMutation::SetFriendStatus(friend_status) => {
-            friend_report.status = friend_status.clone();
-        },
-        FriendReportMutation::SetNumPendingUserRequests(num_pending_user_requests) => {
-            friend_report.num_pending_user_requests = *num_pending_user_requests;
-        },
-        FriendReportMutation::SetOptLastIncomingMoveToken(opt_last_incoming_move_token) => {
-            friend_report.opt_last_incoming_move_token = opt_last_incoming_move_token.clone();
-        },
-        FriendReportMutation::SetLiveness(friend_liveness_report) => {
-            friend_report.liveness = friend_liveness_report.clone();
-        },
-    }
-}
-
-#[allow(unused)]
-pub fn funder_report_mutate<A>(funder_report: &mut FunderReport<A>, mutation: &FunderReportMutation<A>) 
-    -> Result<(), ReportMutateError> 
-where
-    A: Clone,
-{
-
-    match mutation {
-        FunderReportMutation::SetAddress(address) => {
-            funder_report.address = address.clone();
-            Ok(())
-        },
-        FunderReportMutation::AddFriend(add_friend_report) => {
-            let friend_report = FriendReport {
-                remote_address: add_friend_report.address.clone(),
-                name: add_friend_report.name.clone(),
-                sent_local_address: SentLocalAddressReport::NeverSent,
-                opt_last_incoming_move_token: add_friend_report.opt_last_incoming_move_token.clone(),
-                liveness: FriendLivenessReport::Offline,
-                channel_status: add_friend_report.channel_status.clone(),
-                wanted_remote_max_debt: 0,
-                wanted_local_requests_status: RequestsStatusReport::from(&RequestsStatus::Closed),
-                num_pending_responses: 0,
-                num_pending_requests: 0,
-                status: FriendStatusReport::from(&FriendStatus::Disabled),
-                num_pending_user_requests: 0,
-            };
-            if let Some(_) = funder_report.friends.insert(
-                add_friend_report.friend_public_key.clone(), friend_report) {
-
-                Err(ReportMutateError::FriendAlreadyExists)
-            } else {
-                Ok(())
-            }
-        },
-        FunderReportMutation::RemoveFriend(friend_public_key) => {
-            if let None = funder_report.friends.remove(&friend_public_key) {
-                Err(ReportMutateError::FriendDoesNotExist)
-            } else {
-                Ok(())
-            }
-        },
-        FunderReportMutation::FriendReportMutation((friend_public_key, friend_report_mutation)) => {
-            let mut friend = funder_report.friends.get_mut(friend_public_key)
-                .ok_or(ReportMutateError::FriendDoesNotExist)?;
-            friend_report_mutate(&mut friend, friend_report_mutation);
-            Ok(())
-        },
-        FunderReportMutation::SetNumReadyReceipts(num_ready_receipts) => {
-            funder_report.num_ready_receipts = *num_ready_receipts;
-            Ok(())
-        },
-    }
-}
-
-
-// Conversion to index client mutations and state
-// ----------------------------------------------
-
-// This code is used as glue between FunderReport structure and input mutations given to
-// `index_client`. This allows the offst-index-client crate to not depend on the offst-funder
-// crate.
-
-/// Calculate send and receive capacities for a given `friend_report`.
-#[allow(unused)]
-fn calc_friend_capacities<A>(friend_report: &FriendReport<A>) -> (u128, u128) {
-    if friend_report.status == FriendStatusReport::Disabled || 
-        friend_report.liveness == FriendLivenessReport::Offline {
-        return (0, 0);
-    }
-
-    let tc_report = match &friend_report.channel_status {
-        ChannelStatusReport::Inconsistent(_) => return (0,0),
-        ChannelStatusReport::Consistent(tc_report) => tc_report,
-    };
-
-    let balance = &tc_report.balance;
-
-    let send_capacity = if tc_report.requests_status.remote == RequestsStatusReport::Closed {
-        0
-    } else {
-        balance.local_max_debt.saturating_sub(
-            balance.local_pending_debt.checked_sub_signed(balance.balance).unwrap())
-    };
-
-    let recv_capacity = if tc_report.requests_status.local == RequestsStatusReport::Closed {
-        0
-    } else {
-        balance.remote_max_debt.saturating_sub(
-            balance.remote_pending_debt.checked_add_signed(balance.balance).unwrap())
-    };
-
-    (send_capacity, recv_capacity)
-}
-
-#[allow(unused)]
-pub fn funder_report_to_index_client_state<A>(funder_report: &FunderReport<A>) -> IndexClientState 
-where
-    A: Clone,
-{
-    let friends = funder_report.friends
-        .iter()
-        .map(|(friend_public_key, friend_report)| 
-             (friend_public_key.clone(), calc_friend_capacities(friend_report)))
-        .filter(|(_, (send_capacity, recv_capacity))| *send_capacity != 0 || *recv_capacity != 0)
-        .collect::<HashMap<PublicKey,(u128, u128)>>();
-
-    IndexClientState {
-        friends,
-    }
-}
-
-#[allow(unused)]
-pub fn funder_report_mutation_to_index_client_mutation<A>(funder_report: &FunderReport<A>, 
-                                                      funder_report_mutation: &FunderReportMutation<A>) -> Option<IndexMutation> 
-where
-    A: Clone,
-{
-
-    let create_update_friend = |public_key: &PublicKey| {
-        let mut new_funder_report = funder_report.clone();
-        funder_report_mutate(&mut new_funder_report, funder_report_mutation).unwrap();
-        
-        let new_friend_report = new_funder_report.friends
-            .get(public_key)
-            .unwrap(); // We assert that a new friend was added
-
-        let (send_capacity, recv_capacity) = calc_friend_capacities(new_friend_report);
-        let update_friend = UpdateFriend {
-            public_key: public_key.clone(),
-            send_capacity,
-            recv_capacity,
-        };
-        IndexMutation::UpdateFriend(update_friend)
-    };
-
-    match funder_report_mutation {
-        FunderReportMutation::SetAddress(_) | 
-        FunderReportMutation::SetNumReadyReceipts(_) => None,
-        FunderReportMutation::AddFriend(add_friend_report) => 
-            Some(create_update_friend(&add_friend_report.friend_public_key)),
-        FunderReportMutation::RemoveFriend(public_key) => Some(IndexMutation::RemoveFriend(public_key.clone())),
-        FunderReportMutation::FriendReportMutation((public_key, _friend_report_mutation)) =>
-            Some(create_update_friend(&public_key)),
-    }
-}
